+++ Sistema_envios/Sistema_envios/Administracion/Agregar.aspx.cs
@@ -17,6 +17,7 @@ namespace Sistema_envios.Administracion
         {
             if (!Page.IsPostBack)
             {
+                // Carga a el dropdownList estadoPedido los items en el enum EEstadoPedido
                 var itemNames = System.Enum.GetNames(typeof(EEstadoPedido));
                 for (int i = 0; i <= itemNames.Length - 1; i++)
                 {
@@ -26,20 +27,28 @@ namespace Sistema_envios.Administracion
             }
 
         }
-
+        /// <summary>
+        /// Cambia la propiedad Text de fechaEntrega segun la seleccion en Calendario
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         protected void Calendario_SelectionChanged(object sender, EventArgs e)
         {
             fechaEntrega.Text = calendario.SelectedDate.ToShortDateString();
         }
-
+        /// <summary>
+        /// Agrega un nuevi pedido a la BD
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         protected void btnRegistrar_Click(object sender, EventArgs e)
         {
             Pedido nuevoPedido = new Pedido
             {
                 Proveedor = proveedor.Text,
                 FechaEntrega = calendario.SelectedDate.Date,
-                MontoPagado = float.Parse(montoPagado.Text),
-                MontoTotal = float.Parse(montoTotal.Text),
+                MontoPagado = decimal.Parse(montoPagado.Text),
+                MontoTotal = decimal.Parse(montoTotal.Text),
                 EstadoPedido = (EEstadoPedido)Enum.Parse(typeof(EEstadoPedido), estadoPedido.Text)
             };
